requirements/freeze.txt
certifi==2017.4.17
chardet==3.0.4
coverage==4.4.1
decorator==4.0.11
Django==1.11.7
django-bootstrap3==8.2.3
django-cors-headers==2.1.0
django-fe-core==0.1.0
django-oauth-toolkit==1.0.0
djangorestframework==3.6.3
djangorestframework-jwt==1.10.0
drf-nested-routers==0.90.0
factory-boy==2.8.1
Faker==0.7.17
idna==2.5
ipaddress==1.0.18
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
oauthlib==2.0.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.34
Pygments==2.2.0
PyJWT==1.5.2
pytest==3.1.3
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-runner==2.11.1
python-dateutil==2.6.0
pytz==2017.2
requests==2.18.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
